+++ python/lsst/ts/ess/common/sensor/omega_hx85ba.py
@@ -126,7 +126,6 @@ class Hx85baSensor(BaseSensor):
             If a value is missing because the connection to the sensor is
             established mid output, then the value gets replaced by math.nan.
         """
-        self.log.debug("extract_telemetry")
         stripped_line: str = line.strip(self.terminator)
         line_items = stripped_line.split(self.delimiter)
         output = []