+++ python/lsst/ts/ess/common/sensor/temperature_sensor.py
@@ -84,7 +84,6 @@ class TemperatureSensor(BaseSensor):
             or if a channel is missing because the connection to the sensor is
             established mid output, then the value gets replaced by math.nan.
         """
-        self.log.debug("extract_telemetry")
         stripped_line: str = line.strip(self.terminator)
         line_items = stripped_line.split(self.delimiter)
         output = []